extract window effect manager
